Use a staged build and a smarter build.
We break the build into four stages and run each of those in order within the CI script,
using previously built images as the cache source. We also run in a single build step
because the time to allocate a new remote docker server dwarfs the speed up of parallel
execution for test/lint/typehinting.